ADD new HTMLObject and move MarkdownDescription to utils.misc_objs source file
